backend/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
Pillow
python-dotenv
yt-dlp
numpy